[refactor: Refactor flash sale section and adjust import paths]
