+++ src/templates/pages/homepage.template.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import FlashSale from "../sections/flash sale/flashSale.tsx";
+import FlashSale from "../sections/flashSale/flashSale.tsx";
 import Categories from "../sections/categories/categories.tsx";
 import Heading from "../../components/typography/Heading.tsx";
 import "../../components/productCard/ProductCard.css";
